+++ example/lib/di.dart
@@ -6,6 +6,7 @@ import 'api/product_api.dart';
 import 'repository/product_details_repository.dart';
 import 'repository/product_repository.dart';
 import 'repository/product_secret_code_repository.dart';
+import 'repository/visibility_repository.dart';
 import 'service/delete_product_use_case.dart';
 
 /// Simple di helper. In real projects usually GetIt or any other.
@@ -14,6 +15,7 @@ class Di {
   static final productRepository = ProductRepository(_api);
   static final productDetailsRepository = ProductDetailsRepository(_api);
   static final secretCodeRepository = ProductSecretCodeRepository(_api);
+  static final visibilityRepository = VisibilityRepository();
   static final deleteProductUseCase = DeleteProductUseCase(
     api: _api,
     productRepository: productRepository,