+++ lib/src/util/network_bound_resource.dart
@@ -9,13 +9,11 @@ import 'package:rxdart/rxdart.dart';
 import 'package:synchronized/synchronized.dart';
 
 import '../resource.dart';
+import 'cache_duration.dart';
 
 /// Callback to load resource from external source, usually from the network
 /// by [key] ([key] == [_resourceKey])
 typedef FetchCallable<K, V> = Future<V> Function(K key);
-
-/// Callback to resolve cache duration for each cached value.
-typedef CacheDurationResolver<K, V> = Duration Function(K key, V value);
 
 /// Implementation of NetworkBoundResource to follow the single source of truth
 /// principle.
@@ -52,33 +50,39 @@ class NetworkBoundResource<K, V> {
   /// Each time when [NetworkBoundResource.asStream] called the next sequence
   /// is executed:
   ///
-  /// 1. Retrieve cached data from [storage]
-  ///
-  /// 2. If cached data exists then [cacheDurationResolver] is called to resolve
-  /// cache duration.
-  ///
-  /// 3. If cached data is not stale yet - [Resource.success] emits
-  ///
-  /// 4. If cached data is stale - [Resource.loading] emits and
-  /// [fetch] callback executes to fetch fresh data
-  /// and then [Resource.success] emits
+  /// 1. Immediately emits [Resource.loading] state for new subscriber.
+  ///    If [internalCacheEnabled] then loading state contains
+  ///    the last known value.
+  /// 2. Retrieve cached data from [storage].
+  /// 3. If [fetch] callback is not provided then emits Resource.success
+  ///    with cached data and breaks.
+  /// 4. If [fetch] callback is provided and cached data exists
+  ///    then [cacheDurationResolver] is called to resolve cache duration.
+  /// 5. If cached data is not stale yet - emits [Resource.success] with
+  ///    cached data and breaks.
+  /// 5. If cached data is stale - [Resource.loading] emits and
+  ///    [fetch] callback executes to fetch fresh data.
+  /// 6. Regarding the result of [fetch], emits [Resource.success]
+  ///    or [Resource.error]
   ///
   NetworkBoundResource(
     this._resourceKey, {
-    required CacheDurationResolver<K, V> cacheDurationResolver,
     required ResourceStorage<K, V> storage,
+    required CacheDuration<K, V> cacheDuration,
     FetchCallable<K, V>? fetch,
     Logger? logger,
     TimestampProvider timestampProvider = const TimestampProvider(),
+    bool internalCacheEnabled = true,
   })  : _logger = logger,
         _fetch = fetch,
-        _cacheDurationResolver = cacheDurationResolver,
+        _cacheDuration = cacheDuration,
         _storage = storage,
-        _timestampProvider = timestampProvider;
+        _timestampProvider = timestampProvider,
+        _last = _InternalCache(enabled: internalCacheEnabled);
 
   final K _resourceKey;
   final FetchCallable<K, V>? _fetch;
-  final CacheDurationResolver<K, V> _cacheDurationResolver;
+  final CacheDuration<K, V> _cacheDuration;
   final ResourceStorage<K, V> _storage;
   final Logger? _logger;
 
@@ -87,20 +91,21 @@ class NetworkBoundResource<K, V> {
 
   final _subject = PublishSubject<Resource<V>>();
   final _lock = Lock();
+  final _InternalCache _last;
 
   bool _isLoading = false;
   bool _shouldReload = false;
 
-  /// Triggers resource to load (from cache or external if cache is stale)
-  /// and returns hot stream of resource.
+  /// Creates cold (defer) stream of the resource. On subscribe, it triggers
+  /// resource to load from cache or external source ([_fetch] callback)
+  /// if cache is stale.
   ///
   /// Set [forceReload] = true to force resource reloading from external source
   /// even if cache is not stale yet.
-  ///
-  Stream<Resource<V>> asStream({bool forceReload = false}) {
-    _requestLoading(forceReload: forceReload);
-    return _subject.distinct();
-  }
+  Stream<Resource<V>> asStream({bool forceReload = false}) => Rx.defer(() {
+        _requestLoading(forceReload: forceReload);
+        return _subject.startWith(Resource.loading(_last.value)).distinct();
+      });
 
   /// Applies [edit] function to cached value and emit as new success value
   /// If [notifyOnNull] set as true then will emit success(null) in case
@@ -119,13 +124,18 @@ class NetworkBoundResource<K, V> {
           newValue,
           storeTime: cache?.storeTime ?? 0,
         );
-        _subject.add(Resource.success(newValue));
+        _emit(Resource.success(newValue));
       } else if (cache != null) {
         // newValue is null, so we need to remove old one from cache
         await _storage.remove(_resourceKey);
-        if (notifyOnNull) _subject.add(Resource.success(null));
+        if (notifyOnNull) _emit(Resource.success(null));
       }
     });
+  }
+
+  void _emit(Resource<V> resource) {
+    _last.value = resource.data;
+    _subject.add(resource);
   }
 
   /// Returns cached value if exists
@@ -145,20 +155,27 @@ class NetworkBoundResource<K, V> {
   /// Puts new value in cache and emits Resource.success(value)
   Future<void> putValue(V value) => _lock.synchronized(() async {
         await _storage.put(_resourceKey, value);
-        _subject.add(Resource.success(value));
+        _emit(Resource.success(value));
       });
 
   /// Removes resource associated to [_resourceKey] from cache
   Future<void> clearCache() => _storage.remove(_resourceKey);
 
   /// Make cache stale.
-  /// Also triggers resource reloading if [forceReload] is true (by default)
+  /// Also triggers resource reloading if [forceReload] is true (by default).
+  /// If forceReload=true and emitLoading=true then emits loading state firstly.
   /// Returns future that completes after reloading finished with success
   /// or error.
-  Future<void> invalidate([bool forceReload = true]) async {
+  Future<void> invalidate({
+    bool forceReload = true,
+    bool emitLoading = false,
+  }) async {
     // don't clear cache for offline usage, just override store time
     await _overrideStoreTime(0);
     if (forceReload) {
+      if (emitLoading) {
+        _emit(Resource.loading(_last.value));
+      }
       await asStream(forceReload: true)
           .where((event) => event.isNotLoading)
           .first;
@@ -169,9 +186,7 @@ class NetworkBoundResource<K, V> {
   /// New subscriptions will fail after this call.
   Future<void> close() => _subject.close();
 
-  void _requestLoading({
-    bool forceReload = false,
-  }) async {
+  void _requestLoading({bool forceReload = false}) async {
     if (forceReload) {
       _shouldReload = true;
     }
@@ -196,12 +211,7 @@ class NetworkBoundResource<K, V> {
         // No need to perform another requested loading as fetch was not called yet
         _shouldReload = false;
         final cache = await _storage.getOrNull(_resourceKey);
-
-        // Try always starting with loading value
-        // to pass through _subject.distinct()
-        _subject
-          ..add(Resource.loading(cache?.value))
-          ..add(Resource.success(cache?.value));
+        _emit(Resource.success(cache?.value));
       });
 
   Future<void> _loadFromExternal() => _lock.synchronized(() async {
@@ -209,10 +219,9 @@ class NetworkBoundResource<K, V> {
         // get value from cache
         final cache = await _storage.getOrNull(_resourceKey);
 
-        // try always starting with loading value
-        // to pass through _subject.distinct()
-        final resource = Resource.loading(cache?.value);
-        _subject.add(resource);
+        if (_last.value != cache?.value) {
+          _emit(Resource.loading(cache?.value));
+        }
 
         // if cache is stale then need to reload resource from external source
         bool shouldReload =
@@ -224,47 +233,33 @@ class NetworkBoundResource<K, V> {
         if (cache != null && !shouldReload) {
           // There is no external fetch callback, so cache is only source
           // or there is not stale cache and forceReload not requested
-          final resource = Resource.success(cache.value);
-          _subject.add(resource);
+          _emit(Resource.success(cache.value));
           return;
         }
 
         // fetch new value from external source
-        final fetchStream = _fetch!(_resourceKey)
-            .asStream()
-            .asyncMap((data) async {
-              //store new value in the cache before emitting
-              await _storage.put(_resourceKey, data);
-              return data;
-            })
-            .map((data) => Resource.success(data))
-            .onErrorReturnWith((error, trace) {
-              _logger?.trace(
-                  LoggerLevel.error,
-                  'Error loading resource by key [$_resourceKey]',
-                  error,
-                  trace);
-              return Resource.error(
-                'Error loading resource by key [$_resourceKey]',
-                error: error,
-                stackTrace: trace,
-                data: cache?.value,
-              );
-            });
+        final fetchStream =
+            _fetch!(_resourceKey).asStream().asyncMap((data) async {
+          //store new value in the cache before emitting
+          await _storage.put(_resourceKey, data);
+          _last.value = data;
+          return Resource.success(data);
+        }).onErrorReturnWith((error, trace) {
+          _logger?.trace(LoggerLevel.error,
+              'Error loading resource by key [$_resourceKey]', error, trace);
+          return Resource.error(
+            'Error loading resource by key [$_resourceKey]',
+            error: error,
+            stackTrace: trace,
+            data: cache?.value,
+          );
+        });
 
         return _subject.addStream(fetchStream);
       });
 
-  bool _isCacheStale(CacheEntry<V> cache) {
-    if (cache.storeTime <= 0) {
-      // seems like cached time was reset, so resource is stale
-      return true;
-    }
-    final cacheDuration =
-        _cacheDurationResolver(_resourceKey, cache.value).inMilliseconds;
-    final now = _timestampProvider.getTimestamp();
-    return cache.storeTime < now - cacheDuration;
-  }
+  bool _isCacheStale(CacheEntry<V> cache) =>
+      _cacheDuration.isCacheStale(_resourceKey, cache, _timestampProvider);
 
   Future<void> _overrideStoreTime(int storeTime) async {
     _lock.synchronized(() async {
@@ -273,5 +268,22 @@ class NetworkBoundResource<K, V> {
         await _storage.put(_resourceKey, cache.value, storeTime: storeTime);
       }
     });
+  }
+}
+
+/// Internal fast cache for last emitted value.
+/// Can be disabled for security reason to not keep a value in memory
+class _InternalCache<V> {
+  _InternalCache({required this.enabled});
+
+  final bool enabled;
+  V? _cache;
+
+  V? get value => _cache;
+
+  set value(V? newValue) {
+    if (enabled) {
+      _cache = newValue;
+    }
   }
 }